Refactor in BCH
